pyproject.toml
[project]
name = "gateway-service"
version = "0.1.0"
description = ""

requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pydantic-extra-types (>=2.11.0,<3.0.0)",
    "phonenumbers (>=9.0.22,<10.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "ruff (>=0.14.14,<0.15.0)",
    "kirt08-exceptions (>=0.1.0,<0.2.0)",
    "kirt08-tokens (>=0.1.2,<0.2.0)",
    "kirt08-contracts (>=0.2.3,<0.3.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
